Moved macro implementation to individual 2.10/2.11/2.12 versions (needs to be consolidated)
